@@ -8,44 +8,21 @@
   },
   "require": {
     "php": ">=7",
-    "symfony/symfony": "2.7.*",
-    "doctrine/orm": "~2.2,>=2.2.3,<2.5",
-    "doctrine/dbal": "<2.5",
-    "doctrine/doctrine-bundle": "~1.4",
-    "symfony/assetic-bundle": "~2.3",
+    "symfony/symfony": "~3.0",
+    "doctrine/orm": "~2.5",
+    "doctrine/dbal": "~2.5",
     "symfony/monolog-bundle": "~2.4",
-    "sensio/distribution-bundle": "~4.0",
-    "sensio/framework-extra-bundle": "~3.0,>=3.0.2",
-    "incenteev/composer-parameter-handler": "~2.0",
-    "doctrine/migrations": "1.0.*@dev",
-    "doctrine/doctrine-migrations-bundle": "1.0.*"
+    "doctrine/doctrine-bundle": "~1.4",
+    "doctrine/migrations": "dev-master",
+    "doctrine/doctrine-migrations-bundle": "dev-master"
   },
   "require-dev": {
     "phpunit/phpunit": "~4",
-    "squizlabs/php_codesniffer": "~2",
-    "sensio/generator-bundle": "~2.3"
+    "squizlabs/php_codesniffer": "~2"
   },
   "scripts": {
     "post-root-package-install": [
       "SymfonyStandard\\Composer::hookRootPackageInstall"
-    ],
-    "post-install-cmd": [
-      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::removeSymfonyStandardFiles",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
-    ],
-    "post-update-cmd": [
-      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::removeSymfonyStandardFiles",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
     ]
   },
   "config": {
